refactor: Refactor X-Adapter folders structure EX-6298
