+++ packages/x-adapter/src/endpoint-adatpter/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export * from './endpoint-adapter.factory';
