+++ packages/x-adapter/src/endpoint-adatpter/types.ts
@@ -1,5 +1,5 @@
-import { HttpClient, RequestOptions } from './http-client.types';
-import { Mapper } from './mapper.types';
+import { HttpClient, RequestOptions } from '../http-clients/types';
+import { Mapper } from '../mappers/types';
 
 /**
  * A facade containing all the different {@link EndpointAdapter} methods.
