+++ packages/x-adapter/src/http-clients/index.ts
@@ -1,3 +1,4 @@
 export * from './fetch.http-client';
 export * from './beacon.http-client';
+export * from './types';
 export * from './utils';
