+++ packages/x-adapter/src/index.ts
@@ -1,6 +1,5 @@
-export * from './endpoint-adapter.factory';
+export * from './endpoint-adatpter';
 export * from './http-clients';
 export * from './mappers';
 export * from './schemas';
-export * from './types';
 export * from './utils';
