+++ packages/x-adapter/src/mappers/index.ts
@@ -1,2 +1,3 @@
 export * from './identity.mapper';
 export * from './schema-mapper.factory';
+export * from './types';
