+++ packages/x-adapter/src/schemas/__tests__/schemas.types.spec.ts
@@ -1,4 +1,4 @@
-import { Schema } from '../schemas.types';
+import { Schema } from '../types';
 
 describe('schema tests', () => {
   interface Source {
