+++ packages/x-adapter/src/schemas/index.ts
@@ -1,2 +1,2 @@
 export * from './utils';
-export * from './schemas.types';
+export * from './types';
