--- packages/x-adapter/src/types/index.ts
@@ -1,3 +1,0 @@
-export * from './adapter.types';
-export * from './http-client.types';
-export * from './mapper.types';
